exploit: add --dump for sboot dumping mode
In some cases the device's boot partition is still intact and contains
an unknown version of sboot. In order to compile a working shellcode,
one must know the exact addresses of certain functions in sboot. Allow
the user to dump his version of sboot in order to adapt the shellcode
accordingly.